+++ src/components/Forms/Select/Select.spec.js
@@ -1,28 +1,12 @@
 import React from 'react';
 import { render, cleanup, fireEvent, waitForDomChange } from '@testing-library/react';
-import noop from 'lodash/fp/noop';
 import Select from '.';
-import SelectDropDown from './SelectDropDown';
-import SelectOptionGroup from './SelectOptionGroup';
-import SelectOption from './SelectOption';
-import SelectTrigger from './SelectTrigger';
 
 const optionsMock = [
   { label: 'Foo', value: 'foo' },
   { label: 'Bar', value: 'bar' },
   { label: 'Baz', value: 'baz', disabled: true },
 ];
-
-const optionsGroupMock = [
-  {
-    name: 'foo',
-    options: [
-      { label: 'foo', value: 'foo' },
-      { label: 'bar', value: 'bar' },
-    ],
-  },
-];
-
 
 describe('Select component', () => {
   afterEach(() => {
@@ -80,6 +64,13 @@ describe('Select component', () => {
     expect(selectComp).to.exist;
   });
 
+  it('should accept a fluid prop', () => {
+    const { container } = render(<Select fluid options={optionsMock} />);
+    const selectComp = container.querySelector('.bi.bi-select');
+
+    expect(selectComp.classList.contains('fluid')).to.be.true;
+  });
+
   it('if provided, should perform the onChange callback when an option is selected', () => {
     const onChange = sinon.spy();
 
@@ -119,7 +110,6 @@ describe('Select component', () => {
 
     return waitingForDomChangesPromise;
   });
-
 
   it('should keep the dropdown open when selecting an option if \'toggleOnChange\' is set to false', () => {
     const { container } = render(<Select options={optionsMock} value={optionsMock[1].value} toggleOnChange={false} />);
@@ -220,184 +210,5 @@ describe('Select component', () => {
     expect(container.querySelector('.bi-select-element .bi.bi-pill')).to.not.exist;
 
     fireEvent.click(selectComp);
-  });
-});
-
-
-describe('SelectDropDown component', () => {
-  afterEach(() => {
-    sinon.restore();
-    cleanup();
-  });
-
-  it('should render without explode', () => {
-    const { container } = render(
-      <SelectDropDown
-        options={optionsMock}
-        filtrable
-        filterInputPlaceholder="foo"
-        filterNoResultLabel="bar"
-        onChange={noop}
-      />,
-    );
-
-    expect(container).to.exist;
-    expect(container.querySelector('.bi-select-options-dropdown')).to.exist;
-  });
-
-  it('should have default classes', () => {
-    const { container } = render(
-      <SelectDropDown
-        options={optionsMock}
-        filtrable
-        filterInputPlaceholder="foo"
-        filterNoResultLabel="bar"
-        onChange={noop}
-      />,
-    );
-
-    const selectDropDownComp = container.querySelector('div');
-
-    expect(selectDropDownComp.classList.contains('bi-select-options-dropdown')).to.be.true;
-  });
-
-  it('should display a grouped dropdown list', () => {
-    const { container } = render(
-      <SelectDropDown
-        options={optionsGroupMock}
-        hasOptionGroups
-        filtrable
-        filterInputPlaceholder="foo"
-        filterNoResultLabel="bar"
-        onChange={noop}
-        value="foo"
-      />,
-    );
-
-    const selectDropDownComp = container.querySelector('div');
-
-    expect(selectDropDownComp.classList.contains('grouped-opts')).to.be.true;
-  });
-
-  it('should not display a grouped dropdown if hasOptionGroup is undefined ', () => {
-    const { container } = render(
-      <SelectDropDown
-        options={optionsMock}
-        filtrable
-        filterInputPlaceholder="foo"
-        filterNoResultLabel="bar"
-        onChange={noop}
-      />,
-    );
-
-    const selectDropDownComp = container.querySelector('div');
-
-    expect(selectDropDownComp.classList.contains('grouped-opts')).to.be.false;
-  });
-});
-
-describe('SelectOption component', () => {
-  afterEach(() => {
-    sinon.restore();
-    cleanup();
-  });
-
-  it('should render without explode', () => {
-    const { container } = render(<SelectOption option={optionsMock[0]} onClick={noop} />);
-
-    expect(container).to.exist;
-    expect(container.querySelector('.bi-select-opt')).to.exist;
-  });
-
-  it('should have default classes', () => {
-    const { container } = render(<SelectOption option={optionsMock[0]} onClick={noop} />);
-    const selectOptionComp = container.querySelector('div');
-
-    expect(selectOptionComp.classList.contains('bi-select-opt')).to.be.true;
-  });
-
-  it('should perform the onClick callback when an option is selected and yet available', () => {
-    const onClick = sinon.spy();
-    const { container } = render(<SelectOption option={optionsMock[0]} onClick={onClick} />);
-    const selectOptionComp = container.querySelector('.bi-select-opt');
-    fireEvent.click(selectOptionComp);
-    expect(onClick.called).to.equal(true);
-  });
-
-  it('should not perform the onClick callback if an option is disabled', () => {
-    const onClick = sinon.spy();
-    const { container } = render(<SelectOption option={optionsMock[2]} onClick={onClick} />);
-    const selectOptionComp = container.querySelector('.bi-select-opt');
-    fireEvent.click(selectOptionComp);
-    expect(onClick.called).to.equal(false);
-  });
-
-  it('should show the option as selected if its \'selected\' property is set to true', () => {
-    const { container } = render(<SelectOption option={optionsMock[0]} onClick={noop} selected />);
-    const selectOptionComp = container.querySelector('.bi-select-opt');
-
-    expect(selectOptionComp.classList.contains('selected'));
-  });
-});
-
-describe('SelectOptionGroup component', () => {
-  afterEach(() => {
-    sinon.restore();
-    cleanup();
-  });
-
-  it('should render without explode', () => {
-    const { container } = render(<SelectOptionGroup group={optionsGroupMock[0]} value="foo" onClick={noop} />);
-
-    expect(container).to.exist;
-    expect(container.querySelector('.select-group')).to.exist;
-  });
-
-  it('should have default classes', () => {
-    const { container } = render(<SelectOptionGroup group={optionsGroupMock[0]} value="foo" onClick={noop} />);
-    const selectOptionGroupComp = container.querySelector('div');
-
-    expect(selectOptionGroupComp.classList.contains('select-group')).to.be.true;
-  });
-});
-
-describe('SelectTrigger component', () => {
-  afterEach(() => {
-    sinon.restore();
-    cleanup();
-  });
-
-  it('should render without explode', () => {
-    const { container } = render(<SelectTrigger options={optionsMock} />);
-
-    expect(container).to.exist;
-    expect(container.querySelector('.bi-select-element')).to.exist;
-  });
-
-  it('should have default classes', () => {
-    const { container } = render(<SelectTrigger options={optionsMock} />);
-    const selectTriggerComp = container.querySelector('div').querySelector('div');
-
-    expect(selectTriggerComp.classList.contains('bi-select-element')).to.be.true;
-  });
-
-  it('should perform the onClear callback when clicking on the close button', () => {
-    const onClear = sinon.spy();
-    const { container } = render(<SelectTrigger options={optionsMock} onClear={onClear} clearable />);
-    const selectTriggerComp = container.querySelector('.bi-select-element');
-    const closeIcon = selectTriggerComp.querySelector('.sel-clear-x');
-
-    expect(closeIcon).to.exist;
-
-    fireEvent.click(closeIcon);
-
-    expect(onClear.called).to.be.true;
-  });
-
-  it('should possibly show a helping text', () => {
-    render(<SelectTrigger options={optionsMock} helpText="foo" />);
-    const helpTextComp = document.querySelector('.bi.bi-helptext');
-
-    expect(helpTextComp).to.exist;
   });
 });